Move account activities in the account package
Grr, had forgotten that I wanted to move these files.

These five Activities are never started by the main Seadroid
Activities (e.g. BrowserActivity or AccountsActivity).

They are only ever started by the "Authenticator" or by each
other. For that reason I find it more fitting to group them
together within the "account" package.

This patch does just the moving and should therefore have to
further impact.
