+++ src/com/seafile/seadroid2/account/ui/ShibbolethActivity.java
@@ -1,4 +1,4 @@
-package com.seafile.seadroid2.ui.activity;
+package com.seafile.seadroid2.account.ui;
 
 import android.content.Intent;
 import android.os.Bundle;
